structuring and organizing server-side code
